Move REPL pipeline functions to separate module
