+++ src/Juvix/Compiler/Pipeline.hs
@@ -15,10 +15,8 @@ import Juvix.Compiler.Backend qualified as Backend
 import Juvix.Compiler.Backend.C qualified as C
 import Juvix.Compiler.Backend.Geb qualified as Geb
 import Juvix.Compiler.Builtins
-import Juvix.Compiler.Concrete.Data.ParsedInfoTableBuilder qualified as C
 import Juvix.Compiler.Concrete.Data.ParsedInfoTableBuilder qualified as Concrete
 import Juvix.Compiler.Concrete.Data.Scope
-import Juvix.Compiler.Concrete.Data.Scope qualified as Scoper
 import Juvix.Compiler.Concrete.Language
 import Juvix.Compiler.Concrete.Translation.FromParsed qualified as Scoper
 import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver
@@ -42,180 +40,6 @@ import Juvix.Prelude
 type PipelineEff = '[PathResolver, Reader EntryPoint, Files, NameIdGen, Builtins, Error JuvixError, Embed IO]
 
 type TopPipelineEff = '[PathResolver, Reader EntryPoint, Files, NameIdGen, Builtins, State Artifacts, Error JuvixError, Embed IO]
-
-arityCheckExpression ::
-  Members '[Error JuvixError, State Artifacts] r =>
-  ExpressionAtoms 'Parsed ->
-  Sem r Internal.Expression
-arityCheckExpression p = do
-  scopeTable <- gets (^. artifactScopeTable)
-  ( runNameIdGenArtifacts
-      . runBuiltinsArtifacts
-      . runScoperScopeArtifacts
-    )
-    $ Scoper.scopeCheckExpression scopeTable p
-      >>= Abstract.fromConcreteExpression
-      >>= Internal.fromAbstractExpression
-      >>= Internal.arityCheckExpression
-
-openImportToInternal ::
-  Members '[Reader EntryPoint, Error JuvixError, State Artifacts] r =>
-  OpenModule 'Parsed ->
-  Sem r (Maybe Internal.Include)
-openImportToInternal o = do
-  parsedModules <- gets (^. artifactParsing . C.stateModules)
-  ( runNameIdGenArtifacts
-      . runBuiltinsArtifacts
-      . runAbstractInfoTableBuilderArtifacts
-      . runScoperInfoTableBuilderArtifacts
-      . runScoperScopeArtifacts
-      . runStateArtifacts artifactInternalTranslationState
-      . runReaderArtifacts artifactScopeExports
-      . runReader (Scoper.ScopeParameters mempty parsedModules)
-      . runStateArtifacts artifactAbstractModuleCache
-      . runStateArtifacts artifactScoperState
-    )
-    $ do
-      mTopModule <-
-        Scoper.scopeCheckOpenModule o
-          >>= Abstract.fromConcreteOpenImport
-      case mTopModule of
-        Nothing -> return Nothing
-        Just m -> Internal.fromAbstractImport m
-
-importToInternal ::
-  Members '[Reader EntryPoint, Error JuvixError, State Artifacts] r =>
-  Import 'Parsed ->
-  Sem r (Maybe Internal.Include)
-importToInternal i = do
-  parsedModules <- gets (^. artifactParsing . C.stateModules)
-  ( runNameIdGenArtifacts
-      . runBuiltinsArtifacts
-      . runAbstractInfoTableBuilderArtifacts
-      . runScoperInfoTableBuilderArtifacts
-      . runScoperScopeArtifacts
-      . runStateArtifacts artifactInternalTranslationState
-      . runReaderArtifacts artifactScopeExports
-      . runReader (Scoper.ScopeParameters mempty parsedModules)
-      . runStateArtifacts artifactAbstractModuleCache
-      . runStateArtifacts artifactScoperState
-    )
-    $ Scoper.scopeCheckImport i
-      >>= Abstract.fromConcreteImport
-      >>= Internal.fromAbstractImport
-
-importToInternal' ::
-  Members '[Reader EntryPoint, Error JuvixError, State Artifacts] r =>
-  Internal.Include ->
-  Sem r Internal.Include
-importToInternal' = Internal.arityCheckInclude >=> Internal.typeCheckInclude
-
-parseExpression ::
-  Members '[State Artifacts, Error JuvixError] r =>
-  Path Abs File ->
-  Text ->
-  Sem r (ExpressionAtoms 'Parsed)
-parseExpression fp txt =
-  ( runNameIdGenArtifacts
-      . runBuiltinsArtifacts
-  )
-    $ Parser.expressionFromTextSource fp txt
-
-parseReplInput ::
-  Members '[PathResolver, Files, State Artifacts, Error JuvixError] r =>
-  Path Abs File ->
-  Text ->
-  Sem r Parser.ReplInput
-parseReplInput fp txt =
-  ( runNameIdGenArtifacts
-      . runBuiltinsArtifacts
-      . runParserInfoTableBuilderArtifacts
-  )
-    $ Parser.replInputFromTextSource fp txt
-
-inferExpression ::
-  Members '[Error JuvixError, State Artifacts] r =>
-  Path Abs File ->
-  Text ->
-  Sem r Internal.Expression
-inferExpression fp txt = do
-  p <- parseExpression fp txt
-  arityCheckExpression p
-    >>= Internal.inferExpressionType
-
-compileExpression ::
-  Members '[Error JuvixError, State Artifacts] r =>
-  ExpressionAtoms 'Parsed ->
-  Sem r Core.Node
-compileExpression p = do
-  arityCheckExpression p
-    >>= Internal.typeCheckExpression
-    >>= fromInternalExpression
-
-registerImport ::
-  Members '[Error JuvixError, State Artifacts, Reader EntryPoint] r =>
-  Import 'Parsed ->
-  Sem r ()
-registerImport i = do
-  mInclude <- importToInternal i
-  whenJust mInclude (importToInternal' >=> fromInternalInclude)
-
-registerOpenImport ::
-  Members '[Error JuvixError, State Artifacts, Reader EntryPoint] r =>
-  OpenModule 'Parsed ->
-  Sem r ()
-registerOpenImport o = do
-  mInclude <- openImportToInternal o
-  whenJust mInclude (importToInternal' >=> fromInternalInclude)
-
-fromInternalInclude :: Members '[State Artifacts] r => Internal.Include -> Sem r ()
-fromInternalInclude i = do
-  let table = Internal.buildTable [i ^. Internal.includeModule]
-  runReader table
-    . runCoreInfoTableBuilderArtifacts
-    . runFunctionsTableArtifacts
-    . readerTypesTableArtifacts
-    . runReader Core.initIndexTable
-    $ Core.goTopModule (i ^. Internal.includeModule)
-
-fromInternalExpression :: Members '[State Artifacts] r => Internal.Expression -> Sem r Core.Node
-fromInternalExpression exp = do
-  typedTable <- gets (^. artifactInternalTypedTable)
-  runReader typedTable
-    . tmpCoreInfoTableBuilderArtifacts
-    . runFunctionsTableArtifacts
-    . readerTypesTableArtifacts
-    . runReader Core.initIndexTable
-    $ Core.goExpression exp
-
-data ReplPipelineResult
-  = ReplPipelineResultNode Core.Node
-  | ReplPipelineResultImport TopModulePath
-  | ReplPipelineResultOpenImport Name
-
-compileReplInputIO ::
-  Members '[Reader EntryPoint, State Artifacts, Embed IO] r =>
-  Path Abs File ->
-  Text ->
-  Sem r (Either JuvixError ReplPipelineResult)
-compileReplInputIO fp txt =
-  runError
-    . runFilesIO
-    . runPathResolverArtifacts
-    $ do
-      p <- parseReplInput fp txt
-      case p of
-        Parser.ReplExpression e -> ReplPipelineResultNode <$> compileExpression e
-        Parser.ReplImport i -> registerImport i $> ReplPipelineResultImport (i ^. importModule)
-        Parser.ReplOpenImport i -> registerOpenImport i $> ReplPipelineResultOpenImport (i ^. openModuleName)
-
-inferExpressionIO ::
-  Members '[State Artifacts, Embed IO] r =>
-  Path Abs File ->
-  Text ->
-  Sem r (Either JuvixError Internal.Expression)
-inferExpressionIO fp txt =
-  runError (inferExpression fp txt)
 
 --------------------------------------------------------------------------------
 -- Workflows
